creates a serverless application for use in a lab
The app uses the base todo app from vuejs/vue as the frontend, it has been modified to allow the use of a backend api using serverless technologies.
The todo app is licensed using the MIT license.

The API creates four endpoints to complete a basic CRUD API. The create endpoint uses VTL to add a message to a queue. The queue approach is completely superficial but does add another target to run labs against.