Update _index.md
In the Reserved Instance and Savings Plan Example, there was a line of text embedded in the tsql code section.  There should actually be 2 sections of code, one for when there is SP and RI usage both, an one when there is SP usage only.  I have edited this page to separate the 2 code bodies and move the line of text between them outside of the code blocks.