Update 2_budget_ec2actual.md (#646)
Updated Level 100_2_Cost and Usage Governance - Step 4: Budget_EC2_Actual. A step was missing. The "Filter" Option will not be active unless the Budget Scope is changed from the default/recommended Scope Option, which is "All AWS Services (Recommended).
For this step since we are looking to ONLY look at EC2 Instances, we will need to select the "Filter Specific AWS cost dimensions. This will now give us the OPTIONS to select EC2 (or other services as required).

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.