Update PDF link to current version of whitepaper (#808)
* Update PDF link to current version of whitepaper

Previously, link to PDF was to an archived version from July 2020

* Remove link to WAF pdf and change T&C link

This patch removes PDF link to the WAF security to prevent users from landing on deprecated version when the whitepaper is updated. Traffic now directs to the htm. It also updates T&C link to point to AWS Skill Builder per suggestion